Move instruction related structs to their own file
