+++ programs/bpf_loader/src/lib.rs
@@ -49,7 +49,9 @@ use {
     solana_svm_measure::measure::Measure,
     solana_svm_type_overrides::sync::{atomic::Ordering, Arc},
     solana_system_interface::{instruction as system_instruction, MAX_PERMITTED_DATA_LENGTH},
-    solana_transaction_context::{IndexOfAccount, InstructionContext, TransactionContext},
+    solana_transaction_context::{
+        instruction::InstructionContext, IndexOfAccount, TransactionContext,
+    },
     std::{cell::RefCell, mem, rc::Rc},
 };
